# perf: cache `FormMeta` directly