--- a/frappe/desk/form/meta.py
+++ b/frappe/desk/form/meta.py
@@ -36,11 +36,9 @@ def get_meta(doctype, cached=True):
 	# don't cache for developer mode as js files, templates may be edited
 	if cached and not frappe.conf.developer_mode:
 		meta = frappe.cache().hget("form_meta", doctype)
-		if meta:
-			meta = FormMeta(meta)
-		else:
+		if not meta:
 			meta = FormMeta(doctype)
-			frappe.cache().hset("form_meta", doctype, meta.as_dict())
+			frappe.cache().hset("form_meta", doctype, meta)
 	else:
 		meta = FormMeta(doctype)