Similar to https://github.com/frappe/frappe/pull/18164 for for `FormMeta`. This object is much more complex, so can take longer to unpickle.

## 67% faster for Sales Invoice (Redis-only testing)

### Before

```python
In [4]: %timeit get_meta("Sales Invoice")
5.67 ms ± 139 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

### After

```python
In [4]: %timeit get_meta("Sales Invoice")
1.84 ms ± 33 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```